app/requirements.txt
Flask>=3.0.0
openai
ragstack-ai
python-dotenv